Fix the CSP-dead UI, the contaminated build config, and the CI lint
THE APP'S PRIMARY NAVIGATION WAS INVISIBLE.
The shipped CSP is `style-src 'self'` with no nonce or hash, so WebKit drops
every runtime <style> element AND every inline style= attribute. Six widgets
had their styling only in TypeScript, so they rendered completely unstyled:
the six-destination sidebar came out as native grey OS buttons wrapping across
the top of the window, and .native-discord-composer-unreachable -- the
role="alert" chip that warns plaintext is leaving the app -- had no alarm
affordance at all. All six now have real rules in styles.css, verified under
the exact production CSP in webkit2gtk-4.1, the same engine the app embeds.

Nothing caught this because ia-sidebar.test.ts asserted the CSS string appeared
in primarySidebarMarkup()'s OWN SOURCE TEXT. It pinned a CSP-dead string and
reported green the whole time the navigation was destroyed. It now reads
styles.css and was mutation-proved to fail. Same pattern removed from
discord-qa-whitelist-warning.test.ts.

BUILD CONFIG WAS CONTAMINATED BY QA BRANCHES.
- tauri.conf.json's main window pointed at whatsapp-qa.html, so the shipping app
  opened the WhatsApp QA harness instead of OSL. The WhatsApp lane now has its
  own tauri.whatsapp-qa.conf.json, matching how the Signal lane already worked.
- vite.config.ts named rollup inputs, which disables Vite's implicit index.html
  entry -- so dist/ had no index.html and no main-*.js (385 kB). The main UI
  bundle had stopped being built entirely.

CI HAS BEEN RED SINCE 2026-07-22 ON ONE LINT.
clippy::question_mark in crates/ipc/src/commands.rs, plus the rest of a
--workspace --all-targets -D warnings sweep. Where a lint was deliberate
(osl-ratchet-next denies clippy::indexing_slicing because a panic on
attacker-controlled input is a DoS) the deny is kept for production code and
exempted only in test modules.

Also: migration 0037 renumbered off a duplicate 0036 (it reads a table the
other 0036 creates, and only ran in the right order by alphabetical luck);
ipc state_reload now surfaces a successful wrong-key quarantine instead of
only a failed rename; three stale source-audit assertions corrected; and the
osl-hub Tauri command surface moved into a lib module so its tests are
actually compiled -- 9 of them were built by nothing and run by nothing.

Suites: root workspace 1789/1789, osl-hub lib 971/971, osl-hub full 1037/1038,
osl-hub-ui 1207/1207, keyserver-cf 494/494 + 126/126, keyserver 76/76.

The one remaining failure is documented in place at the Err branch in
broker.rs: a replayed notice whose single-use wrapped key is gone is never
retired and never counted as deferred, so it occupies one of the 32 rows
admitted per pair forever. Retiring it is a product decision, entangled with
whether a replay may re-emit a Received ack (it must not).

+++ crates/ipc/src/main_password.rs
@@ -923,10 +923,8 @@ pub fn run_inactivity_auto_lock_timer_for_state(
     if get_file_storage_key().is_none() {
         return InactivityAutoLockOutcome::AlreadyLocked;
     }
-    let timer = keystore::InactivityTimer::with_last_activity(
-        INACTIVITY_AUTO_LOCK_SECONDS as u64,
-        last_activity,
-    );
+    let timer =
+        keystore::InactivityTimer::with_last_activity(INACTIVITY_AUTO_LOCK_SECONDS, last_activity);
     if timer.should_reprompt_at(now) {
         lock_main_password_session(state);
         InactivityAutoLockOutcome::Locked
@@ -1893,7 +1891,7 @@ mod password_policy_tests {
         assert_eq!(
             super::run_inactivity_auto_lock_timer_for_state(
                 &state,
-                now - std::time::Duration::from_secs((INACTIVITY_AUTO_LOCK_SECONDS - 1) as u64),
+                now - std::time::Duration::from_secs(INACTIVITY_AUTO_LOCK_SECONDS - 1),
                 now,
             ),
             InactivityAutoLockOutcome::StillUnlocked
@@ -1905,7 +1903,7 @@ mod password_policy_tests {
         assert_eq!(
             super::run_inactivity_auto_lock_timer_for_state(
                 &state,
-                now - std::time::Duration::from_secs(INACTIVITY_AUTO_LOCK_SECONDS as u64),
+                now - std::time::Duration::from_secs(INACTIVITY_AUTO_LOCK_SECONDS),
                 now,
             ),
             InactivityAutoLockOutcome::Locked
@@ -1920,7 +1918,7 @@ mod password_policy_tests {
         assert_eq!(
             super::run_inactivity_auto_lock_timer_for_state(
                 &state,
-                now - std::time::Duration::from_secs((INACTIVITY_AUTO_LOCK_SECONDS * 2) as u64),
+                now - std::time::Duration::from_secs(INACTIVITY_AUTO_LOCK_SECONDS * 2),
                 now,
             ),
             InactivityAutoLockOutcome::AlreadyLocked
